+++ R/provider-deepseek.R
@@ -29,7 +29,6 @@ chat_deepseek <- function(
   api_key = deepseek_key(),
   model = NULL,
   params = NULL,
-  seed = NULL,
   api_args = list(),
   echo = NULL,
   api_headers = character()
@@ -44,7 +43,6 @@ chat_deepseek <- function(
     base_url = base_url,
     model = model,
     params = params,
-    seed = seed,
     extra_args = api_args,
     api_key = api_key,
     extra_headers = api_headers