+++ tests/testthat/test-provider-azure.R
@@ -70,30 +70,6 @@ test_that("Azure request headers are generated correctly", {
   )
   req <- chat_request(p, FALSE, list(turn))
   expect_snapshot(str(req_get_headers(req, "reveal")))
-
-  # Token.
-  p <- ProviderAzureOpenAI(
-    name = "Azure",
-    base_url = base_url,
-    model = deployment_id,
-    api_version = "2024-06-01",
-    api_key = "",
-    credentials = default_azure_credentials("", "token")
-  )
-  req <- chat_request(p, FALSE, list(turn))
-  expect_snapshot(str(req_get_headers(req, "reveal")))
-
-  # Both.
-  p <- ProviderAzureOpenAI(
-    name = "Azure",
-    base_url = base_url,
-    model = deployment_id,
-    api_version = "2024-06-01",
-    api_key = "key",
-    credentials = default_azure_credentials("key", "token")
-  )
-  req <- chat_request(p, FALSE, list(turn))
-  expect_snapshot(str(req_get_headers(req, "reveal")))
 })
 
 test_that("service principal authentication requests look correct", {